pyproject.toml
[tool.poetry]
name = "Joinee"
version = "0.1.0"
description = ""
authors = ["wsa-2002"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.10"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
asyncpg = "^0.29.0"
databases = "^0.8.0"
python-dotenv = "^1.0.0"
starlette-context = "^0.3.5"
PyJWT = "^2.6.0"
passlib = "1.7.4"
argon2-cffi = "^23.1.0"
python-multipart = "^0.0.6"
python-json-logger = "2.0.7"
PyYAML = "6.0.1"
aioredis = "^2.0.1"
google-cloud-storage = "^2.13.0"
aiosmtplib = "^3.0.1"
pydantic = {extras = ["email"], version = "^2.4.2"}
Authlib = "^1.2.1"
httpx = "^0.25.1"
starlette-session = "^0.4.3"
google-api-python-client = "^2.108.0"
requests = "2.31.0"
google-cloud-logging = "^3.8.0"
googlemaps = "^4.10.0"

[tool.poetry.group.dev.dependencies]
responses = "^0.23.3"
coverage = "^7.3.2"
freezegun = "^1.2.2"
pycodestyle = "^2.11.1"
pytest = "^7.4.3"
parameterized = "^0.9.0"
pre-commit = "^3.5.0"
ipython = "^8.17.2"
isort = "^5.12.0"
add-trailing-comma = "^3.1.0"

[tool.isort]
src_paths = ["app", "tests"]
split_on_trailing_comma = true


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
